Allow cli flags to set logically False values
During testing, `--disable-refresh False` failed to override the
default True value. This fix changes the logic so that properties
are overwritten as long as an argument is not None. Before it only
overwrote a property if an argument value was logically True. This
bug effects every command line argument that overrides a property.